nsfw-semantic-search-api/requirements.txt
flask
flask-cors
torch
transformers
Pillow
numpy
requests
faiss-cpu
sentence-transformers
opencv-python-headless
